backend/requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
SQLAlchemy>=2.0.0
Flask-Migrate>=4.0.5
alembic>=1.13.2
Flask-Mail>=0.9.1
Flask-Bcrypt>=1.0.1
python-dotenv>=1.0.1
itsdangerous>=2.2.0
Jinja2>=3.1.4
Werkzeug>=3.0.3
gunicorn>=21.2.0
requests>=2.32.3
qrcode[pil]>=7.4
Pillow>=10.4.0
Flask-Login>=0.6.3
Flask-Cors>=4.0.0
Flask-Login>=0.6.3
Flask-Cors>=4.0.0
